+++ Source/NoteCreator.h
@@ -47,7 +47,7 @@ public:
 
    void SetEnabled(bool enabled) override { mEnabled = enabled; }
 
-   void OnPulse(double time, float velocity, int flags) override;
+   void OnPulse(double time, double velocity, int flags) override;
 
    void CheckboxUpdated(Checkbox* checkbox, double time) override;
    void ButtonClicked(ClickButton* button, double time) override;
