+++ Source/PulseHocket.cpp
@@ -105,7 +105,7 @@ void PulseHocket::AdjustHeight()
    mHeight = height;
 }
 
-void PulseHocket::OnPulse(double time, float velocity, int flags)
+void PulseHocket::OnPulse(double time, double velocity, int flags)
 {
    ComputeSliders(0);
