+++ Source/PulseTrain.cpp
@@ -130,12 +130,12 @@ void PulseTrain::OnTimeEvent(double time)
    Step(time, 1, 0);
 }
 
-void PulseTrain::OnPulse(double time, float velocity, int flags)
+void PulseTrain::OnPulse(double time, double velocity, int flags)
 {
    mStep = 0;
 }
 
-void PulseTrain::Step(double time, float velocity, int flags)
+void PulseTrain::Step(double time, double velocity, int flags)
 {
    if (!mEnabled)
       return;
@@ -149,7 +149,7 @@ void PulseTrain::Step(double time, float velocity, int flags)
 
    if (mStep < mLength)
    {
-      float v = mVels[mStep] * velocity;
+      double v = mVels[mStep] * velocity;
 
       int new_flags = 0;
       if (mResetOnStart && mStep == 0)
