+++ Source/Ramper.h
@@ -55,7 +55,7 @@ public:
    void OnTransportAdvanced(double amount) override;
 
    //IPulseReceiver
-   void OnPulse(double time, float velocity, int flags) override;
+   void OnPulse(double time, double velocity, int flags) override;
 
    void DropdownUpdated(DropdownList* list, int oldVal, double time) override {}
    void ButtonClicked(ClickButton* button, double time) override;
